+++ src/test/java/goorm_runner/backend/post/application/post/PostServiceIntegrationTest.java
@@ -1,10 +1,8 @@
-package goorm_runner.backend.post.application;
+package goorm_runner.backend.post.application.post;
 
 import goorm_runner.backend.global.ErrorCode;
 import goorm_runner.backend.member.domain.Member;
 import goorm_runner.backend.member.domain.MemberRepository;
-import goorm_runner.backend.post.application.post.PostReadService;
-import goorm_runner.backend.post.application.post.PostService;
 import goorm_runner.backend.post.application.post.exception.PostException;
 import goorm_runner.backend.post.domain.PostRepository;
 import goorm_runner.backend.post.domain.model.Category;
